requirements.txt
h5py
matplotlib
numpy
pandas
Pillow
torch
torchinfo
torchvision
tqdm
transformers
wandb
